Host implementation done
Host implementation done

ripped out predecessor un-erasure

more simplification and cleanup

more cleanup of host implementation
