+++ core/src/impl/Kokkos_GraphNodeCustomization.hpp
@@ -1,4 +1,3 @@
-
 /*
 //@HEADER
 // ************************************************************************
@@ -43,57 +42,57 @@
 //@HEADER
 */
 
-#include <gtest/gtest.h>
+#ifndef KOKKOS_IMPL_KOKKOS_GRAPHNODECUSTOMIZATION_HPP
+#define KOKKOS_IMPL_KOKKOS_GRAPHNODECUSTOMIZATION_HPP
 
-#include <Kokkos_Core.hpp>
+#include <Kokkos_Macros.hpp>
+#include <Kokkos_Core_fwd.hpp>
+#include <Kokkos_Graph_fwd.hpp>
+#include <impl/Kokkos_GraphImpl_fwd.hpp>
 
-#include <default/TestDefaultDeviceType_Category.hpp>
+namespace Kokkos {
+namespace Impl {
 
-#include <Kokkos_Graph.hpp>
+// Customizable for backends
+template <class ExecutionSpace, class Kernel, class PredecessorRef>
+struct GraphNodeBackendDetailsBeforeTypeErasure {
+ protected:
+  //----------------------------------------------------------------------------
+  // <editor-fold desc="ctors, destructor, and assignment"> {{{2
 
-namespace Test {
+  // Required constructors in customizations:
+  GraphNodeBackendDetailsBeforeTypeErasure(ExecutionSpace const&, Kernel&,
+                                           PredecessorRef const&) noexcept {}
+  GraphNodeBackendDetailsBeforeTypeErasure(
+      ExecutionSpace const&, _graph_node_is_root_ctor_tag,
+      GraphNodeBackendSpecificDetails<ExecutionSpace>&
+      this_as_details) noexcept {}
 
-TEST(defaultdevicetype, development_test) {
-  Kokkos::View<int> count{"graph_kernel_count"};
-  Kokkos::View<int> bugs{"graph_kernel_bugs"};
-  const auto graph = Kokkos::Experimental::create_graph([=](auto builder) {
-    auto root = builder.get_root();
 
-    auto f1 = root.then_parallel_for(
-        Kokkos::RangePolicy<TEST_EXECSPACE>(0, 1), KOKKOS_LAMBDA(long) {
-          bugs() += int(count() != 0);
-          count()++;
-        });
-    auto f2 = f1.then_parallel_for(
-        Kokkos::RangePolicy<TEST_EXECSPACE>(0, 1), KOKKOS_LAMBDA(long) {
-          bugs() += int(count() < 1 || count() > 2);
-          count()++;
-        });
-    auto f3 = f1.then_parallel_for(
-        Kokkos::RangePolicy<TEST_EXECSPACE>(0, 1), KOKKOS_LAMBDA(long) {
-          bugs() += int(count() < 1 || count() > 2);
-          count()++;
-        });
-    builder.when_all(f2, f3).then_parallel_for(
-        Kokkos::RangePolicy<TEST_EXECSPACE>(0, 1), KOKKOS_LAMBDA(long) {
-          bugs() += int(count() != 3);
-          count()++;
-        });
-  });
+  // Not copyable or movable at the concept level, so the default
+  // implementation shouldn't be either.
+  GraphNodeBackendDetailsBeforeTypeErasure() noexcept = delete;
 
-  for(int i = 0; i < 2; ++i) {
-    Kokkos::deep_copy(graph.get_execution_space(), count, 0);
-    Kokkos::deep_copy(graph.get_execution_space(), bugs, 0);
-    graph.submit();
-    auto count_host =
-        Kokkos::create_mirror_view_and_copy(graph.get_execution_space(), count);
-    auto bugs_host =
-        Kokkos::create_mirror_view_and_copy(graph.get_execution_space(), bugs);
-    graph.get_execution_space().fence();
+  GraphNodeBackendDetailsBeforeTypeErasure(
+      GraphNodeBackendDetailsBeforeTypeErasure const&) = delete;
 
-    ASSERT_EQ(count_host(), 4);
-    ASSERT_EQ(bugs_host(), 0);
-  }
-}
+  GraphNodeBackendDetailsBeforeTypeErasure(
+      GraphNodeBackendDetailsBeforeTypeErasure&&) noexcept = delete;
 
-}  // namespace Test
+  GraphNodeBackendDetailsBeforeTypeErasure& operator   =(
+      GraphNodeBackendDetailsBeforeTypeErasure const&) = delete;
+
+  GraphNodeBackendDetailsBeforeTypeErasure& operator       =(
+      GraphNodeBackendDetailsBeforeTypeErasure&&) noexcept = delete;
+
+  ~GraphNodeBackendDetailsBeforeTypeErasure() = default;
+
+  // </editor-fold> end ctors, destructor, and assignment }}}2
+  //----------------------------------------------------------------------------
+};
+
+
+} // end namespace Impl
+} // end namespace Kokkos
+
+#endif  // KOKKOS_KOKKOS_GRAPHNODECUSTOMIZATION_HPP
